+++ discord/client.py
@@ -2267,7 +2267,7 @@ class Client:
         if not isinstance(application_command, _valid_application_command_types):
             raise TypeError(f'application_command must derive from SlashCommand, MessageCommand, or UserCommand')
 
-        if application_command.__application_command_guild_ids__ is not None and self.application_command_guild_ids is not None:
+        if application_command.__application_command_guild_ids__ is None and self.application_command_guild_ids is not None:
             application_command.__application_command_guild_ids__ = self.application_command_guild_ids
             application_command.__application_command_global_command__ = False