+++ CHANGELOG.rst
@@ -147,3 +147,10 @@ February 13, 2022
 - Added the ``spammer`` user flag
 - Changed ``Guild.mfa_level`` to the newly added ``MFALevel`` enum
 - Added the ``str(x)`` method to the ``Locale`` enum
+
+February 14, 2022
+-----------------
+
+- Added the ``slowmode_delay`` keyword argument to ``TextChannel.create_thread`` and ``Message.create_thread``
+- Added the ``reason`` keyword argument to ``Message.create_thread``
+- Added the attachment application command option type